cloud function/requirements.txt
google-cloud-compute==1.14.1
google-cloud-storage==1.43.0